refactor(dashboard): extract pure layout-store helpers from useCatalogLayout
Move the deterministic half of useCatalogLayout (viz.kind size
constraints, geometry normalisation, pack-seed building, saved-layout
reconciliation, add placement, storage IO with injected Storage) into
utils/layoutStore.js so the add/persist/rehydrate cycle is testable
under node --test without a React harness. The hook keeps only the
React state wiring; its public surface is unchanged
(sizeConstraintsForKpi re-exported).

Behaviour preserved, with one deliberate correction folded into the
move: defaultSizeForKpi now treats choropleth-map like map (8x6 default
instead of falling through to the generic 6x6 chart default).
