+++ digit-ui-esbuild/products/dashboard/src/hooks/useCatalogLayout.js
@@ -1,12 +1,14 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import {
-  GRID_COLS,
-  UNIFORM_CHART_SIZE_CONSTRAINTS,
-  MAP_SIZE_CONSTRAINTS,
-  FULL_WIDTH_TABLE_GRID,
-  findFirstOpenPosition,
-} from "../constants/layoutConfig";
 import { compactVertically } from "../utils/gridGeometry";
+import {
+  LEGACY_STORAGE_KEY,
+  buildSeedLayout,
+  resolveInitialLayout,
+  addItemToLayout,
+  readSavedLayout,
+  persistLayout,
+  sizeConstraintsForKpi,
+} from "../utils/layoutStore";
 
 /**
  * useCatalogLayout — the catalog-world layout hook (kpiId-keyed).
@@ -16,119 +18,32 @@ import { compactVertically } from "../utils/gridGeometry";
  * inverted dashboard renders straight from the MDMS catalog (snake_case kpiIds)
  * and seeds its grid from the DashboardPack layout, so it gets a fresh, far
  * smaller hook that speaks kpiIds and reuses only the pure geometry/collision
- * helpers (GRID_COLS, findFirstOpenPosition, compactVertically, swapOnDrop).
+ * helpers.
+ *
+ * All pure logic (constraints, seed/saved reconciliation, add placement,
+ * storage IO) lives in utils/layoutStore.js so the add/persist/rehydrate cycle
+ * is unit-testable; this hook owns only the React state around it.
  *
  * Inputs:
  *   kpis        — { [kpiId]: def }  (catalog map; used for size-constraint by viz.kind)
  *   packLayout  — [{ kpiId, x, y, w, h }]  (DashboardPack defaultLayout — the seed)
  *
- * Returns the same interaction surface AdminDashboard's grid expects:
- *   { layout, onDragStop, onResizeStop, onLayoutChange, resetLayout,
- *     removeWidgetFromLayout, addKpiToLayout, visibleLayoutIds }
+ * Returns the interaction surface AdminDashboard's grid expects:
+ *   { layout, onLayoutChange, resetLayout, removeWidgetFromLayout,
+ *     addKpiToLayout, visibleLayoutIds }
  * where every layout item.i is a kpiId.
  */
 
-const STORAGE_KEY = "ccrs.dashboard.catalog-layout.v1";
+export { sizeConstraintsForKpi } from "../utils/layoutStore";
 
-const CARD_KINDS = new Set([
-  "number-tile-delta",
-  "number-tile",
-  "scalar",
-  "number-tile-sparkline",
-  "sparkline-card",
-]);
+const STORAGE_KEY = LEGACY_STORAGE_KEY;
 
-const KPI_CARD_CONSTRAINTS = { minW: 2, minH: 2, maxW: 6, maxH: 3 };
-const LIST_CONSTRAINTS = { minW: 3, minH: 4, maxW: 12, maxH: 12 };
-
-/** Map a tile's viz.kind to its grid size constraints (the single id-space seam). */
-export function sizeConstraintsForKpi(kpiId, kpis) {
-  const kind = kpis?.[kpiId]?.viz?.kind;
-  if (CARD_KINDS.has(kind)) return KPI_CARD_CONSTRAINTS;
-  switch (kind) {
-    case "map":
-    case "choropleth-map":
-      return MAP_SIZE_CONSTRAINTS;
-    case "sla-risk-table":
-    case "table":
-    case "data-table":
-      return {
-        minW: FULL_WIDTH_TABLE_GRID.minW,
-        minH: FULL_WIDTH_TABLE_GRID.minH,
-        maxW: FULL_WIDTH_TABLE_GRID.maxW,
-        maxH: FULL_WIDTH_TABLE_GRID.maxH,
-      };
-    case "rankedList":
-    case "dow":
-      return LIST_CONSTRAINTS;
-    default:
-      return UNIFORM_CHART_SIZE_CONSTRAINTS; // bar / stacked-bar / horizontal-bar / line / pie
-  }
-}
-
-/** Default size for a freshly-added tile, by kind. */
-function defaultSizeForKpi(kpiId, kpis) {
-  const c = sizeConstraintsForKpi(kpiId, kpis);
-  const kind = kpis?.[kpiId]?.viz?.kind;
-  if (CARD_KINDS.has(kind)) return { w: 2, h: 2 };
-  if (kind === "map") return { w: 8, h: 6 };
-  if (kind === "sla-risk-table" || kind === "table" || kind === "data-table")
-    return { w: 12, h: 5 };
-  return { w: Math.max(c.minW, 6), h: Math.max(c.minH, 6) };
-}
-
-function clampNum(v, min, max, fallback) {
-  const n = Number(v);
-  if (!Number.isFinite(n)) return fallback;
-  return Math.min(max, Math.max(min, n));
-}
-
-/**
- * Normalise a layout item to finite, in-bounds geometry. A malformed MDMS pack
- * or stale localStorage entry can carry NaN / out-of-range w/h/x/y; RGL throws or
- * produces impossible resize bounds on those, so clamp every field to the tile's
- * viz.kind constraints and the grid width. Item must already carry `i` (kpiId).
- */
-function normalizeItem(item, kpis) {
-  const c = sizeConstraintsForKpi(item.i, kpis);
-  const w = clampNum(item.w, c.minW, c.maxW, c.minW);
-  const h = clampNum(item.h, c.minH, c.maxH, c.minH);
-  const x = Math.min(clampNum(item.x, 0, GRID_COLS - 1, 0), GRID_COLS - w);
-  const y = clampNum(item.y, 0, Number.MAX_SAFE_INTEGER, 0);
-  return { i: item.i, x, y, w, h, ...c };
-}
-
-/** Seed layout from the pack, normalised, kpiId-keyed. */
-function buildSeedLayout(packLayout, kpis) {
-  return (packLayout || [])
-    .filter((item) => kpis[item.kpiId])
-    .map((item) =>
-      normalizeItem({ i: item.kpiId, x: item.x, y: item.y, w: item.w, h: item.h }, kpis)
-    );
-}
-
-/**
- * Read the saved layout. Returns `null` ONLY when there is no stored layout (key
- * absent / unparseable); an intentionally-empty array (user cleared every tile)
- * is returned as `[]` so the seed does not re-add the removed tiles on reload.
- */
 function readSaved() {
-  try {
-    const raw = window.localStorage?.getItem(STORAGE_KEY);
-    if (raw == null) return null;
-    const parsed = JSON.parse(raw);
-    return Array.isArray(parsed) ? parsed : null;
-  } catch {
-    return null;
-  }
+  return readSavedLayout(window.localStorage, STORAGE_KEY);
 }
 
 function persist(layout) {
-  try {
-    window.localStorage?.setItem(STORAGE_KEY, JSON.stringify(layout));
-  } catch {
-    /* ignore quota/serialisation errors — layout is non-critical state */
-  }
+  persistLayout(window.localStorage, STORAGE_KEY, layout);
 }
 
 export function useCatalogLayout(kpis, packLayout) {
@@ -145,12 +60,7 @@ export function useCatalogLayout(kpis, packLayout) {
   // or a layout reset, not auto-injected.
   useEffect(() => {
     if (!seed.length) return;
-    const saved = readSaved();
-    const source = saved !== null ? saved : seed;
-    const reconciled = source
-      .filter((item) => kpis[item.i])
-      .map((item) => normalizeItem(item, kpis));
-    setLayout(reconciled);
+    setLayout(resolveInitialLayout(readSaved(), seed, kpis));
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [seed]);
 
@@ -201,16 +111,9 @@ export function useCatalogLayout(kpis, packLayout) {
 
   const addKpiToLayout = useCallback(
     (kpiId, position) => {
-      if (!kpis[kpiId]) return;
       setLayout((prev) => {
-        if (prev.some((item) => item.i === kpiId)) return prev; // no duplicates
-        const c = sizeConstraintsForKpi(kpiId, kpis);
-        const { w, h } = defaultSizeForKpi(kpiId, kpis);
-        const pos = position || findFirstOpenPosition(prev, w, h, GRID_COLS);
-        const next = compactVertically([
-          ...prev,
-          { i: kpiId, x: pos.x, y: pos.y, w, h, ...c },
-        ]);
+        const next = addItemToLayout(prev, kpiId, kpis, position);
+        if (next === prev) return prev; // unknown kpi or already placed
         persist(next);
         return next;
       });
